+++ packages/element/src/mutateElement.ts
@@ -131,7 +131,10 @@ export const mutateElement = <TElement extends Mutable<ExcalidrawElement>>(
     typeof updates.height !== "undefined" ||
     typeof updates.width !== "undefined" ||
     typeof fileId != "undefined" ||
-    typeof points !== "undefined"
+    typeof points !== "undefined" ||
+    typeof updates.strokeStyle !== "undefined" ||
+    typeof updates.strokeWidth !== "undefined" ||
+    typeof updates.roundness !== "undefined"
   ) {
     ShapeCache.delete(element);
   }